+++ api-user/src/main/java/com/wellmeet/restaurant/RestaurantService.java
@@ -4,8 +4,8 @@ import com.wellmeet.common.util.DistanceCalculator;
 import com.wellmeet.domain.member.FavoriteRestaurantDomainService;
 import com.wellmeet.domain.restaurant.RestaurantDomainService;
 import com.wellmeet.domain.restaurant.availabledate.entity.AvailableDate;
-import com.wellmeet.domain.restaurant.entity.BoundingBox;
 import com.wellmeet.domain.restaurant.entity.Restaurant;
+import com.wellmeet.domain.restaurant.model.BoundingBox;
 import com.wellmeet.restaurant.dto.AvailableDateResponse;
 import com.wellmeet.restaurant.dto.NearbyRestaurantResponse;
 import com.wellmeet.restaurant.dto.RepresentativeMenuResponse;
