+++ domain/src/main/java/com/wellmeet/domain/restaurant/RestaurantDomainService.java
@@ -4,10 +4,10 @@ import com.wellmeet.domain.exception.DomainErrorCode;
 import com.wellmeet.domain.exception.WellMeetDomainException;
 import com.wellmeet.domain.restaurant.availabledate.AvailableDateDomainService;
 import com.wellmeet.domain.restaurant.availabledate.entity.AvailableDate;
-import com.wellmeet.domain.restaurant.entity.BoundingBox;
 import com.wellmeet.domain.restaurant.entity.Restaurant;
 import com.wellmeet.domain.restaurant.menu.MenuDomainService;
 import com.wellmeet.domain.restaurant.menu.entity.Menu;
+import com.wellmeet.domain.restaurant.model.BoundingBox;
 import com.wellmeet.domain.restaurant.repository.RestaurantRepository;
 import com.wellmeet.domain.restaurant.review.ReviewDomainService;
 import com.wellmeet.domain.restaurant.review.entity.Review;
